LayoutTests/imported/w3c/web-platform-tests/css/motion/parsing/offset-path-parsing-valid-expected.txt


PASS e.style['offset-path'] = "none" should set the property value
FAIL e.style['offset-path'] = "ray(0deg)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['offset-path'] = "ray(0rad closest-side)" should set the property value assert_equals: serialization should be canonical expected "ray(0rad)" but got "ray(0rad closest-side)"
PASS e.style['offset-path'] = "ray(0.25turn closest-corner contain)" should set the property value
PASS e.style['offset-path'] = "ray(200grad farthest-side)" should set the property value
PASS e.style['offset-path'] = "ray(270deg farthest-corner contain)" should set the property value
PASS e.style['offset-path'] = "ray(-720deg sides)" should set the property value
PASS e.style['offset-path'] = "ray(calc(180deg - 45deg) farthest-side)" should set the property value
FAIL e.style['offset-path'] = "ray(0deg at 100px 100px)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['offset-path'] = "ray(0deg sides at center center)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['offset-path'] = "ray(0deg at center center sides)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['offset-path'] = "ray(0deg at center center contain)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['offset-path'] = "ray(at 10px 10px 0deg contain)" should set the property value assert_not_equals: property should be set got disallowed value ""
PASS e.style['offset-path'] = "path(\"m 20 0 h -100\")" should set the property value
PASS e.style['offset-path'] = "path(\"M 0 0 L 100 100 M 100 200 L 200 200 Z L 300 300 Z\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10 20 q 30 60 40 50 q 100 70 90 80\")" should set the property value
PASS e.style['offset-path'] = "path(\"M 0 0 L 100 100 m 0 100 l 100 0 Z l 160 20 Z\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10 20 l 20 30 Z l 50 60 Z m 70 80 l 90 60 Z t 70 120\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10 170 h 90 v 30 m 0 0 s 1 2 3 4 z c 9 8 7 6 5 4\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10 20 a 10 20 30 1 0 40 50 a 110 120 30 1 1 140 50\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10 20   l 20 30   Z   l 50 60   Z   m 70 80   l 90 60   Z   t 70 120\")" should set the property value
PASS e.style['offset-path'] = "path(\"m 10.0 170.0 h 90.00 v 30.00 m 0 0 s 1 2 3 4 z c 9 8 7 6 5 4\")" should set the property value
PASS e.style['offset-path'] = "  path(  \"m 10 20 a 10 20 30 1 0 40 50 a 110 120 30 1 1 140 50\"  )  " should set the property value
PASS e.style['offset-path'] = "url(\"http://www.example.com/index.html#polyline1\")" should set the property value
FAIL e.style['offset-path'] = "circle(100px)" should set the property value assert_equals: serialization should be canonical expected "circle(100px)" but got "circle(100px at 50% 50%)"
PASS e.style['offset-path'] = "circle(100px at 50% 50%)" should set the property value
PASS e.style['offset-path'] = "border-box" should set the property value
PASS e.style['offset-path'] = "inset(10% 20% 30% 40%) border-box" should set the property value
FAIL e.style['offset-path'] = "fill-box ellipse(50% 60%)" should set the property value assert_equals: serialization should be canonical expected "ellipse(50% 60%) fill-box" but got "fill-box ellipse(50% 60% at 50% 50%)"
FAIL e.style['offset-path'] = "content-box ellipse(50% 60% at 50% 50%)" should set the property value assert_equals: serialization should be canonical expected "ellipse(50% 60% at 50% 50%) content-box" but got "content-box ellipse(50% 60% at 50% 50%)"
